style: 코드 정리. viewcontroller에 필요한 코드만 들어가도록 수정
