+++ frontend/YellowLine/YellowLine.xcodeproj/xcuserdata/belltiger.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -14,8 +14,8 @@
             filePath = "YellowLine/CameraSession.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "117"
-            endingLineNumber = "117"
+            startingLineNumber = "143"
+            endingLineNumber = "143"
             landmarkName = "setupOutput()"
             landmarkType = "7">
          </BreakpointContent>
